Working in sim-run, only missing testing on FPGA
