--- software/src/AES256.c
@@ -1,268 +1,0 @@
-#include "versat_accel.h"
-#include "unitConfiguration.h"
-
-#include "iob-uart.h"
-#include "iob_str.h"
-#include "printf.h"
-
-#include "stdlib.h"
-#include "stdint.h"
-
-#define ARRAY_SIZE(ARR) (sizeof(ARR)/sizeof(ARR[0]))
-
-// AES Sizes (bytes)
-#define AES_BLK_SIZE (16)
-#define AES_KEY_SIZE (32)
-
-const uint32_t sbox[256] = {
-   0x63, 0x7c, 0x77, 0x7b, 0xf2, 0x6b, 0x6f, 0xc5, 0x30, 0x01, 0x67, 0x2b, 0xfe, 0xd7, 0xab, 0x76,
-   0xca, 0x82, 0xc9, 0x7d, 0xfa, 0x59, 0x47, 0xf0, 0xad, 0xd4, 0xa2, 0xaf, 0x9c, 0xa4, 0x72, 0xc0,
-   0xb7, 0xfd, 0x93, 0x26, 0x36, 0x3f, 0xf7, 0xcc, 0x34, 0xa5, 0xe5, 0xf1, 0x71, 0xd8, 0x31, 0x15,
-   0x04, 0xc7, 0x23, 0xc3, 0x18, 0x96, 0x05, 0x9a, 0x07, 0x12, 0x80, 0xe2, 0xeb, 0x27, 0xb2, 0x75,
-   0x09, 0x83, 0x2c, 0x1a, 0x1b, 0x6e, 0x5a, 0xa0, 0x52, 0x3b, 0xd6, 0xb3, 0x29, 0xe3, 0x2f, 0x84,
-   0x53, 0xd1, 0x00, 0xed, 0x20, 0xfc, 0xb1, 0x5b, 0x6a, 0xcb, 0xbe, 0x39, 0x4a, 0x4c, 0x58, 0xcf,
-   0xd0, 0xef, 0xaa, 0xfb, 0x43, 0x4d, 0x33, 0x85, 0x45, 0xf9, 0x02, 0x7f, 0x50, 0x3c, 0x9f, 0xa8,
-   0x51, 0xa3, 0x40, 0x8f, 0x92, 0x9d, 0x38, 0xf5, 0xbc, 0xb6, 0xda, 0x21, 0x10, 0xff, 0xf3, 0xd2,
-   0xcd, 0x0c, 0x13, 0xec, 0x5f, 0x97, 0x44, 0x17, 0xc4, 0xa7, 0x7e, 0x3d, 0x64, 0x5d, 0x19, 0x73,
-   0x60, 0x81, 0x4f, 0xdc, 0x22, 0x2a, 0x90, 0x88, 0x46, 0xee, 0xb8, 0x14, 0xde, 0x5e, 0x0b, 0xdb,
-   0xe0, 0x32, 0x3a, 0x0a, 0x49, 0x06, 0x24, 0x5c, 0xc2, 0xd3, 0xac, 0x62, 0x91, 0x95, 0xe4, 0x79,
-   0xe7, 0xc8, 0x37, 0x6d, 0x8d, 0xd5, 0x4e, 0xa9, 0x6c, 0x56, 0xf4, 0xea, 0x65, 0x7a, 0xae, 0x08,
-   0xba, 0x78, 0x25, 0x2e, 0x1c, 0xa6, 0xb4, 0xc6, 0xe8, 0xdd, 0x74, 0x1f, 0x4b, 0xbd, 0x8b, 0x8a,
-   0x70, 0x3e, 0xb5, 0x66, 0x48, 0x03, 0xf6, 0x0e, 0x61, 0x35, 0x57, 0xb9, 0x86, 0xc1, 0x1d, 0x9e,
-   0xe1, 0xf8, 0x98, 0x11, 0x69, 0xd9, 0x8e, 0x94, 0x9b, 0x1e, 0x87, 0xe9, 0xce, 0x55, 0x28, 0xdf,
-   0x8c, 0xa1, 0x89, 0x0d, 0xbf, 0xe6, 0x42, 0x68, 0x41, 0x99, 0x2d, 0x0f, 0xb0, 0x54, 0xbb, 0x16 };
-
-const uint32_t mul2[] = {
-   0x00,0x02,0x04,0x06,0x08,0x0a,0x0c,0x0e,0x10,0x12,0x14,0x16,0x18,0x1a,0x1c,0x1e,
-   0x20,0x22,0x24,0x26,0x28,0x2a,0x2c,0x2e,0x30,0x32,0x34,0x36,0x38,0x3a,0x3c,0x3e,
-   0x40,0x42,0x44,0x46,0x48,0x4a,0x4c,0x4e,0x50,0x52,0x54,0x56,0x58,0x5a,0x5c,0x5e,
-   0x60,0x62,0x64,0x66,0x68,0x6a,0x6c,0x6e,0x70,0x72,0x74,0x76,0x78,0x7a,0x7c,0x7e,
-   0x80,0x82,0x84,0x86,0x88,0x8a,0x8c,0x8e,0x90,0x92,0x94,0x96,0x98,0x9a,0x9c,0x9e,
-   0xa0,0xa2,0xa4,0xa6,0xa8,0xaa,0xac,0xae,0xb0,0xb2,0xb4,0xb6,0xb8,0xba,0xbc,0xbe,
-   0xc0,0xc2,0xc4,0xc6,0xc8,0xca,0xcc,0xce,0xd0,0xd2,0xd4,0xd6,0xd8,0xda,0xdc,0xde,
-   0xe0,0xe2,0xe4,0xe6,0xe8,0xea,0xec,0xee,0xf0,0xf2,0xf4,0xf6,0xf8,0xfa,0xfc,0xfe,
-   0x1b,0x19,0x1f,0x1d,0x13,0x11,0x17,0x15,0x0b,0x09,0x0f,0x0d,0x03,0x01,0x07,0x05,
-   0x3b,0x39,0x3f,0x3d,0x33,0x31,0x37,0x35,0x2b,0x29,0x2f,0x2d,0x23,0x21,0x27,0x25,
-   0x5b,0x59,0x5f,0x5d,0x53,0x51,0x57,0x55,0x4b,0x49,0x4f,0x4d,0x43,0x41,0x47,0x45,
-   0x7b,0x79,0x7f,0x7d,0x73,0x71,0x77,0x75,0x6b,0x69,0x6f,0x6d,0x63,0x61,0x67,0x65,
-   0x9b,0x99,0x9f,0x9d,0x93,0x91,0x97,0x95,0x8b,0x89,0x8f,0x8d,0x83,0x81,0x87,0x85,
-   0xbb,0xb9,0xbf,0xbd,0xb3,0xb1,0xb7,0xb5,0xab,0xa9,0xaf,0xad,0xa3,0xa1,0xa7,0xa5,
-   0xdb,0xd9,0xdf,0xdd,0xd3,0xd1,0xd7,0xd5,0xcb,0xc9,0xcf,0xcd,0xc3,0xc1,0xc7,0xc5,
-   0xfb,0xf9,0xff,0xfd,0xf3,0xf1,0xf7,0xf5,0xeb,0xe9,0xef,0xed,0xe3,0xe1,0xe7,0xe5
-};
-
-const uint32_t mul3[] = {
-   0x00,0x03,0x06,0x05,0x0c,0x0f,0x0a,0x09,0x18,0x1b,0x1e,0x1d,0x14,0x17,0x12,0x11,
-   0x30,0x33,0x36,0x35,0x3c,0x3f,0x3a,0x39,0x28,0x2b,0x2e,0x2d,0x24,0x27,0x22,0x21,
-   0x60,0x63,0x66,0x65,0x6c,0x6f,0x6a,0x69,0x78,0x7b,0x7e,0x7d,0x74,0x77,0x72,0x71,
-   0x50,0x53,0x56,0x55,0x5c,0x5f,0x5a,0x59,0x48,0x4b,0x4e,0x4d,0x44,0x47,0x42,0x41,
-   0xc0,0xc3,0xc6,0xc5,0xcc,0xcf,0xca,0xc9,0xd8,0xdb,0xde,0xdd,0xd4,0xd7,0xd2,0xd1,
-   0xf0,0xf3,0xf6,0xf5,0xfc,0xff,0xfa,0xf9,0xe8,0xeb,0xee,0xed,0xe4,0xe7,0xe2,0xe1,
-   0xa0,0xa3,0xa6,0xa5,0xac,0xaf,0xaa,0xa9,0xb8,0xbb,0xbe,0xbd,0xb4,0xb7,0xb2,0xb1,
-   0x90,0x93,0x96,0x95,0x9c,0x9f,0x9a,0x99,0x88,0x8b,0x8e,0x8d,0x84,0x87,0x82,0x81,
-   0x9b,0x98,0x9d,0x9e,0x97,0x94,0x91,0x92,0x83,0x80,0x85,0x86,0x8f,0x8c,0x89,0x8a,
-   0xab,0xa8,0xad,0xae,0xa7,0xa4,0xa1,0xa2,0xb3,0xb0,0xb5,0xb6,0xbf,0xbc,0xb9,0xba,
-   0xfb,0xf8,0xfd,0xfe,0xf7,0xf4,0xf1,0xf2,0xe3,0xe0,0xe5,0xe6,0xef,0xec,0xe9,0xea,
-   0xcb,0xc8,0xcd,0xce,0xc7,0xc4,0xc1,0xc2,0xd3,0xd0,0xd5,0xd6,0xdf,0xdc,0xd9,0xda,
-   0x5b,0x58,0x5d,0x5e,0x57,0x54,0x51,0x52,0x43,0x40,0x45,0x46,0x4f,0x4c,0x49,0x4a,
-   0x6b,0x68,0x6d,0x6e,0x67,0x64,0x61,0x62,0x73,0x70,0x75,0x76,0x7f,0x7c,0x79,0x7a,
-   0x3b,0x38,0x3d,0x3e,0x37,0x34,0x31,0x32,0x23,0x20,0x25,0x26,0x2f,0x2c,0x29,0x2a,
-   0x0b,0x08,0x0d,0x0e,0x07,0x04,0x01,0x02,0x13,0x10,0x15,0x16,0x1f,0x1c,0x19,0x1a
-};
-
-void FillLookupTable(LookupTableAddr addr){
-   VersatMemoryCopy(addr.addr,(int*) sbox,256 * sizeof(int));
-}
-
-void FillMul(LookupTableAddr addr,const uint32_t* mem,int size){
-   VersatMemoryCopy(addr.addr,(int*) mem,size * sizeof(int));
-}
-
-void FillRow(DoRowAddr addr){
-   FillMul(addr.mul2_0,mul2,ARRAY_SIZE(mul2));
-   FillMul(addr.mul2_1,mul2,ARRAY_SIZE(mul2));
-   FillMul(addr.mul3_0,mul3,ARRAY_SIZE(mul3));
-   FillMul(addr.mul3_1,mul3,ARRAY_SIZE(mul3));
-}
-
-void FillFirstLineKey(FirstLineKeyAddr addr){
-   FillLookupTable(addr.b0);
-   FillLookupTable(addr.b1);
-}
-
-void FillFourthLineKey(FourthLineKeyAddr addr){
-   FillLookupTable(addr.b0);
-   FillLookupTable(addr.b1);
-}
-
-void FillMixColumns(MixColumnsAddr addr){
-   FillRow(addr.d0);
-   FillRow(addr.d1);
-   FillRow(addr.d2);
-   FillRow(addr.d3);
-}
-
-void FillSBox(SBoxAddr addr){
-   FillLookupTable(addr.s0);
-   FillLookupTable(addr.s1);
-   FillLookupTable(addr.s2);
-   FillLookupTable(addr.s3);
-   FillLookupTable(addr.s4);
-   FillLookupTable(addr.s5);
-   FillLookupTable(addr.s6);
-   FillLookupTable(addr.s7);
-}
-
-void FillKeySchedule(KeySchedule256Addr addr){
-   FillFirstLineKey(addr.s);
-   FillFourthLineKey(addr.q);
-}
-
-void FillMainRound(MainRoundAddr addr){
-   FillSBox(addr.subBytes);
-   FillMixColumns(addr.mixColumns);
-}
-
-void FillRoundPairAndKey(RoundPairAndKeyAddr addr){
-   FillMainRound(addr.round1);
-   FillMainRound(addr.round2);
-   FillKeySchedule(addr.key);
-}
-
-void byte_to_int(uint8_t *in, int *out, int size) {
-   int i = 0;
-   for(i=0; i<size; i++) {
-      out[i] = (int) in[i];
-   }
-   return;
-}
-
-void int_to_byte(int *in, uint8_t *out, int size) {
-   int i = 0;
-   for(i=0; i<size; i++) {
-      out[i] = (uint8_t) (in[i] & 0x0FF);
-   }
-   return;
-}
-
-void Versat_init_AES() {
-   AES256WithIterative_SimpleAddr addr = ACCELERATOR_TOP_ADDR_INIT;
-
-   FillRoundPairAndKey(addr.simple.mk0.roundPairAndKey);
-   FillMainRound(addr.simple.round0);
-   FillKeySchedule(addr.simple.key6);
-   FillSBox(addr.simple.subBytes);
-
-   ACCEL_TOP_simple_rcon0_constant = 0x01;
-   ACCEL_TOP_simple_rcon1_constant = 0x02;
-   ACCEL_TOP_simple_rcon2_constant = 0x04;
-   ACCEL_TOP_simple_rcon3_constant = 0x08;
-   ACCEL_TOP_simple_rcon4_constant = 0x10;
-   ACCEL_TOP_simple_rcon5_constant = 0x20;
-   ACCEL_TOP_simple_rcon6_constant = 0x40;
-
-   return;
-}
-
-void VersatAES(uint8_t *result, uint8_t *cypher, uint8_t *key) {
-   int cypher_int[AES_BLK_SIZE] = {0};
-   int key_int[AES_KEY_SIZE] = {0};
-   int result_int[AES_BLK_SIZE] = {0};
-
-   byte_to_int(cypher, cypher_int, AES_BLK_SIZE);
-   byte_to_int(key, key_int, AES_KEY_SIZE);
-
-   int i = 0;
-   for(i = 0; i < AES_BLK_SIZE; i++){
-      SimpleInputStart[i] = cypher_int[i];
-   }
-   for(i = 0; i < AES_KEY_SIZE; i++){
-      SimpleInputStart[i+AES_BLK_SIZE] = key_int[i];
-   }
-
-   RunAccelerator(1);
-
-   for(i = 0; i < AES_BLK_SIZE; i++){
-      result_int[i] = SimpleOutputStart[i];
-   }
-
-   int_to_byte(result_int, result, AES_BLK_SIZE);
-
-   return;
-}
-
-static char GetHexadecimalChar(int value){
-  if(value < 10){
-    return '0' + value;
-  } else{
-    return 'a' + (value - 10);
-  }
-}
-
-char* GetHexadecimal(const char* text,char* buffer, int str_size){
-   int i;
-   for(i = 0; i< str_size; i++){
-      int ch = (int) ((unsigned char) text[i]);
-
-      buffer[i*2] = GetHexadecimalChar(ch / 16);
-      buffer[i*2+1] = GetHexadecimalChar(ch % 16);
-   }
-   buffer[(i+1)*2] = '\0';
-
-   return buffer;
-}
-
-static char HexToInt(char ch){
-   if('0' <= ch && ch <= '9'){
-      return (ch - '0');
-   } else if('a' <= ch && ch <= 'f'){
-      return ch - 'a' + 10;
-   } else if('A' <= ch && ch <= 'F'){
-      return ch - 'A' + 10;
-   } else {
-      printf("Error, invalid character inside hex string:%c",ch);
-      return 0;
-   }
-}
-
-// Make sure that buffer is capable of storing the whole thing. Returns number of bytes inserted
-int HexStringToHex(char* buffer,const char* str){
-   int inserted = 0;
-   for(int i = 0; ; i += 2){
-      char upper = str[i];
-      char lower = str[i+1];
-
-      if(upper == '\0' || lower == '\0'){
-         if(upper != '\0') printf("Warning: HexString was not divisible by 2\n");
-         break;
-      }   
-
-      buffer[inserted++] = HexToInt(upper) * 16 + HexToInt(lower);
-   }
-
-   return inserted;
-}
-
-void AES256Test(){
-   uint8_t key[128] = {};
-   uint8_t plain[128] = {};
-
-   int keyIndex = HexStringToHex((char*) key,"cc22da787f375711c76302bef0979d8eddf842829c2b99ef3dd04e23e54cc24b");
-   int plainIndex = HexStringToHex((char*) plain,"ccc62c6b0a09a671d64456818db29a4d");
-
-   printf("keyIndex: %d\n",keyIndex);
-   printf("plainIndex: %d\n",plainIndex);
-
-   uint8_t result[AES_BLK_SIZE] = {};
-
-   Versat_init_AES();
-
-   VersatAES(result,plain,key);
-
-   char buffer[2048];
-   GetHexadecimal((char*) result,buffer, AES_BLK_SIZE);
-
-   printf("df8634ca02b13a125b786e1dce90658b\n");
-   printf("%s\n",buffer);
-}
-
