+++ software/src/unitConfiguration.h
@@ -1,4 +1,5 @@
-#include "versat_accel.h"
+#ifndef INCLUDED_UNIT_CONFIGURATION
+#define INCLUDED_UNIT_CONFIGURATION
 
 void IntSet(volatile void* buffer,int value,int byteSize){
    volatile int* asInt = (int*) buffer;
@@ -20,9 +21,6 @@ void ConfigureSimpleVRead(VReadConfig* inst, int numberItems,int* memory){
    inst->pingPong = 1;
    inst->ext_addr = (iptr) memory;
    inst->length = numberItems * sizeof(int);
-   //inst->iterA = 1;
-   //inst->dutyA = numberItems;
-   //inst->int_addr = 0;
 
    // B - versat side
    inst->iterB = 1;
@@ -30,71 +28,17 @@ void ConfigureSimpleVRead(VReadConfig* inst, int numberItems,int* memory){
    inst->perB = numberItems;
    inst->dutyB = 1;
 }
-
-void ConfigureLeftSideMatrixVRead(VReadConfig* inst, int iterations){
-   IntSet(inst,0,sizeof(VReadConfig));
-
-   int numberItems = iterations * iterations;
-
-   inst->incrA = 1;
-   //inst->iterA = 1;
-   inst->perA = numberItems;
-   //inst->dutyA = numberItems;
-   //inst->int_addr = 0;
-   inst->pingPong = 0;
-   inst->length = numberItems * sizeof(int);
-
-   inst->iterB = iterations;
-   inst->perB = iterations;
-   inst->dutyB = iterations;
-   inst->startB = 0;
-   inst->shiftB = -iterations;
-   inst->incrB = 1;
-   inst->reverseB = 0;
-   inst->iter2B = 1;
-   inst->per2B = iterations;
-   inst->shift2B = 0;
-   inst->incr2B = iterations;
-}
-
-void ConfigureRightSideMatrixVRead(VReadConfig* inst, int iterations){
-   IntSet(inst,0,sizeof(VReadConfig));
-
-   int numberItems = iterations * iterations;
-
-   inst->incrA = 1;
-   //inst->iterA = 1;
-   inst->perA = numberItems;
-   //inst->dutyA = numberItems;
-   //inst->int_addr = 0;
-   inst->pingPong = 0;
-   inst->length = numberItems * sizeof(int);
-
-   inst->iterB = iterations;
-   inst->perB = iterations;
-   inst->dutyB = iterations;
-   inst->startB = 0;
-   inst->shiftB = -(iterations * iterations - 1);
-   inst->incrB = iterations;
-   inst->reverseB = 0;
-   inst->iter2B = 1;
-   inst->per2B = iterations;
-   inst->shift2B = 0;
-   inst->incr2B = 0;
-}
 #endif
 
 #ifdef VERSAT_DEFINED_VWrite
+// Not being used for now.
 void ConfigureSimpleVWrite(VWriteConfig* inst, int numberItems,int* memory){
    IntSet(inst,0,sizeof(VWriteConfig));
 
    // Write side
    inst->incrA = 1;
-   inst->iterA = 1;
    inst->perA = numberItems;
-   inst->dutyA = numberItems;
-   inst->size = 8;
-   inst->int_addr = 0;
+   //inst->int_addr = 0;
    inst->pingPong = 1;
    inst->length = numberItems * sizeof(int);
    inst->ext_addr = (iptr) memory;
@@ -105,60 +49,10 @@ void ConfigureSimpleVWrite(VWriteConfig* inst, int numberItems,int* memory){
    inst->dutyB = 1;
    inst->incrB = 1;
 }
-
-void ConfigureMatrixVWrite(VWriteConfig* inst,int amountOfData){
-   IntSet(inst,0,sizeof(VWriteConfig));
-
-   inst->incrA = 1;
-   inst->iterA = 1;
-   inst->perA = amountOfData;
-   inst->dutyA = amountOfData;
-   inst->size = 8;
-   inst->int_addr = 0;
-   inst->pingPong = 0;
-   inst->length = amountOfData * sizeof(int);
-
-   inst->iterB = amountOfData;
-   inst->perB = 4;
-   inst->dutyB = 1;
-   inst->incrB = 1;
-}
 #endif
 
 #ifdef VERSAT_DEFINED_Mem
-void ConfigureLeftSideMatrix(MemConfig* inst,int iterations){
-   IntSet(inst,0,sizeof(MemConfig));
-
-   inst->iterA = iterations;
-   inst->perA = iterations;
-   inst->dutyA = iterations;
-   inst->startA = 0;
-   inst->shiftA = -iterations;
-   inst->incrA = 1;
-   inst->reverseA = 0;
-   inst->iter2A = 1;
-   inst->per2A = iterations;
-   inst->shift2A = 0;
-   inst->incr2A = iterations;
-}
-
-void ConfigureRightSideMatrix(MemConfig* inst, int iterations){
-   IntSet(inst,0,sizeof(MemConfig));
-
-   inst->iterA = iterations;
-   inst->perA = iterations;
-   inst->dutyA = iterations;
-   inst->startA = 0;
-   inst->shiftA = -(iterations * iterations - 1);
-   inst->incrA = iterations;
-   inst->reverseA = 0;
-   inst->iter2A = 1;
-   inst->per2A = iterations;
-   inst->shift2A = 0;
-   inst->incr2A = 0;
-}
-
-void ConfigureMemoryLinear(MemConfig* inst, int amountOfData, int start){
+void ConfigureSimpleMemoryWithStart(MemConfig* inst, int amountOfData, int start){
    IntSet(inst,0,sizeof(MemConfig));
 
    inst->iterA = 1;
@@ -168,7 +62,12 @@ void ConfigureMemoryLinear(MemConfig* inst, int amountOfData, int start){
    inst->startA = start;
 }
 
-void ConfigureMemoryLinearOut(MemConfig* inst, int amountOfData){
+void ConfigureSimpleMemoryAndCopyData(MemConfig* inst, int amountOfData, int start,MemAddr addr,int* data){
+   ConfigureSimpleMemoryWithStart(inst,amountOfData,start);
+   VersatMemoryCopy(addr.addr,data,amountOfData * sizeof(int));
+}
+
+void ConfigureSimpleMemory(MemConfig* inst, int amountOfData){
    IntSet(inst,0,sizeof(MemConfig));
 
    inst->iterA = 1;
@@ -178,21 +77,15 @@ void ConfigureMemoryLinearOut(MemConfig* inst, int amountOfData){
    inst->in0_wr = 1;
 }
 
-void ConfigureMemoryReceive(MemConfig* inst, int amountOfData,int interdataDelay){
+void ConfigureMemoryReceive(MemConfig* inst, int amountOfData){
    IntSet(inst,0,sizeof(MemConfig));
 
-   inst->iterA = amountOfData;
-   inst->perA = interdataDelay;
-   inst->dutyA = 1;
-   inst->startA = 0;
-   inst->shiftA = 0;
+   inst->iterA = 1;
+   inst->perA = amountOfData;
+   inst->dutyA = amountOfData;
    inst->incrA = 1;
    inst->in0_wr = 1;
-   inst->reverseA = 0;
-   inst->iter2A = 0;
-   inst->per2A = 0;
-   inst->shift2A = 0;
-   inst->incr2A = 0;
 }
 #endif
 
+#endif // INCLUDED_UNIT_CONFIGURATION
