+++ software/sw_build.mk
@@ -71,9 +71,7 @@ IOB_SOC_OPENCRYPTOLINUX_LFLAGS=-Wl,-Bstatic,-T,$(TEMPLATE_LDS),--strip-debug
 IOB_SOC_OPENCRYPTOLINUX_FW_SRC=src/iob_soc_opencryptolinux_firmware.S
 IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/iob_soc_opencryptolinux_firmware.c
 IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/printf.c
-#IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/iob_str.c
-#IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/SHA.c
-#IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/AES256.c
+IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/SHA_AES.c
 
 # PERIPHERAL SOURCES
 IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=$(wildcard src/iob-*.c)
